Fix: emit,listening => gateway 통합
