+++ src/channels/channels.module.ts
@@ -6,19 +6,19 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Channel } from './entities/channel.entity';
 import { ChannelRelation } from './entities/channel-relation.entity';
 import { ChannelInvitation } from './entities/channel-invitation.entity';
-import { ChannelsEmitGateway } from './channels-emit.gateway';
-import { CommonsModule } from 'src/commons/commons.module';
-import { NotificationsModule } from 'src/notifications/notifications.module';
+import { ChannelGateway } from './channels.gateway';
+import { UserRelationModule } from 'src/user-relation/user-relation.module';
+import { SocketConnectionModule } from 'src/socket-connection/socket-connection.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([Channel, ChannelRelation, ChannelInvitation]),
-    NotificationsModule,
     UsersModule,
-    CommonsModule,
+    UserRelationModule,
+    SocketConnectionModule,
   ],
   controllers: [ChannelsController],
-  providers: [ChannelsService, ChannelsEmitGateway],
+  providers: [ChannelsService, ChannelGateway],
   exports: [ChannelsService],
 })
 export class ChannelsModule {}
