+++ src/channels/channels.service.ts
@@ -1,10 +1,11 @@
 import { User } from './../users/user.entity';
-
 import {
   BadRequestException,
   ForbiddenException,
+  Inject,
   Injectable,
   NotFoundException,
+  forwardRef,
 } from '@nestjs/common';
 import { Channel, ChannelType } from './entities/channel.entity';
 import { ChannelRelation } from './entities/channel-relation.entity';
@@ -13,8 +14,7 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { ChannelDto } from './dto/channel.dto';
 import * as bcrypt from 'bcrypt';
 import { ChannelInvitation, InvitationStatus } from './entities/channel-invitation.entity';
-import { NotificationsEmitGateway } from 'src/notifications/notifications-emit.gateway';
-import { ChannelsEmitGateway } from './channels-emit.gateway';
+import { ChannelGateway } from './channels.gateway';
 
 @Injectable()
 export class ChannelsService {
@@ -25,8 +25,8 @@ export class ChannelsService {
     private channelRelationRepository: Repository<ChannelRelation>,
     @InjectRepository(ChannelInvitation)
     private channelInvitationRepository: Repository<ChannelInvitation>,
-    private notificationEmitGateway: NotificationsEmitGateway,
-    private channelsEmitGateway: ChannelsEmitGateway,
+    @Inject(forwardRef(() => ChannelGateway))
+    private channelGateway: ChannelGateway,
   ) {}
 
   async createChannel(owner: User, channelDto: ChannelDto): Promise<Channel> {
@@ -131,7 +131,7 @@ export class ChannelsService {
       }
     }
     // 소켓으로 channelRoom에서 나가고 채널 유저들에게 전파
-    await this.channelsEmitGateway.leaveChannelRoom(user.id, channelId);
+    await this.channelGateway.leaveChannelRoom(user.id, channelId);
   }
 
   private async transferOwnership(earliestOwnerRelation: ChannelRelation): Promise<void> {
@@ -192,7 +192,7 @@ export class ChannelsService {
       // ban 전에 kick 처리
       bannedUserRelation.isBanned = true;
       await this.channelRelationRepository.save(bannedUserRelation);
-      await this.channelsEmitGateway.leaveChannelRoom(bannedUserId, channelId);
+      await this.channelGateway.leaveChannelRoom(bannedUserId, channelId);
     } else {
       throw new ForbiddenException('관리자는 다른 관리자나 소유자를 ban할 수 없습니다!');
     }
@@ -240,7 +240,7 @@ export class ChannelsService {
     ) {
       // kick 실시간 소켓 처리
       await this.channelRelationRepository.remove(kickedUserRelation);
-      await this.channelsEmitGateway.leaveChannelRoom(kickedUserId, channelId);
+      await this.channelGateway.leaveChannelRoom(kickedUserId, channelId);
     } else {
       throw new ForbiddenException('관리자는 다른 관리자나 소유자를 kick할 수 없습니다!');
     }
@@ -273,7 +273,7 @@ export class ChannelsService {
 
     requestedUserRelation.isAdmin = updateData.isAdmin;
     await this.channelRelationRepository.save(requestedUserRelation);
-    await this.channelsEmitGateway.emitChannelMemberUpdate(channelId, requestedUserRelation);
+    await this.channelGateway.emitChannelMemberUpdate(channelId, requestedUserRelation);
   }
 
   async changeOwner(channelId: number, currentOwnerId: number, successorId: number): Promise<void> {
@@ -307,8 +307,8 @@ export class ChannelsService {
     successorRelation.isAdmin = true;
 
     await this.channelRelationRepository.save([currentOwnerRelation, successorRelation]);
-    this.channelsEmitGateway.emitChannelMemberUpdate(channelId, currentOwnerRelation);
-    this.channelsEmitGateway.emitChannelMemberUpdate(channelId, successorRelation);
+    this.channelGateway.emitChannelMemberUpdate(channelId, currentOwnerRelation);
+    this.channelGateway.emitChannelMemberUpdate(channelId, successorRelation);
   }
 
   async inviteUser(
@@ -359,7 +359,7 @@ export class ChannelsService {
         status: InvitationStatus.Waiting,
       });
       await this.channelInvitationRepository.save(existingInvitation);
-      await this.notificationEmitGateway.notiChannelInvite(existingInvitation);
+      await this.channelGateway.notiChannelInvite(existingInvitation);
       return existingInvitation;
     }
     // 유저가 이전에 초대되었지만 refuse했던 경우 status를 waiting으로 수정
@@ -371,7 +371,7 @@ export class ChannelsService {
         status: InvitationStatus.Waiting,
       });
       const savedInvitation = await this.channelInvitationRepository.save(newInvitation);
-      await this.notificationEmitGateway.notiChannelInvite(savedInvitation);
+      await this.channelGateway.notiChannelInvite(savedInvitation);
       return savedInvitation;
     }
     // 존재하는 초대가 없거나 이전에 초대를 수락했으면 초대 기록 남기기
@@ -473,7 +473,7 @@ export class ChannelsService {
       user,
     });
     const savedRelation = await this.channelRelationRepository.save(newRelation);
-    await this.channelsEmitGateway.joinChannelRoom(savedRelation);
+    await this.channelGateway.joinChannelRoom(savedRelation);
   }
 
   async findOneChannelRelation(channelId: number, userId: number): Promise<ChannelRelation> {
