+++ src/channels/dto/channel-message-emit.dto.ts
@@ -1,0 +1,14 @@
+import { Expose } from 'class-transformer';
+import { ChannelDto } from './channel.dto';
+import { ShowUserOverviewDto } from 'src/users/dtos/show-user-overview.dto';
+
+export class ChannelMessageEmitDto {
+  @Expose()
+  channel: ChannelDto;
+
+  @Expose()
+  sender: ShowUserOverviewDto;
+
+  @Expose()
+  content: string;
+}
