+++ src/channels/dto/channel-message-receive.dto.ts
@@ -1,0 +1,9 @@
+import { IsNumber, IsString } from 'class-validator';
+
+export class ChannelMessageReceiveDto {
+  @IsNumber()
+  channelId: number;
+
+  @IsString()
+  content: string;
+}
